* Fix bug when back to GamePlayScene from GameOverScene

+++ Cocos2dx/SpaceShooter/Classes/GamePlayScene.h
@@ -14,7 +14,7 @@ private:
 	SpaceShooter* m_spaceShip;
 	Sprite* m_background;
 	Label* m_lbScore;
-	long score;
+	int score;
 	int count;
 public:
